requirements.txt
# Core web and utility libraries
flask>=2.3.0
flask-socketio>=5.3.0
psutil>=5.9.0
requests>=2.31.0

# Google Generative AI library
google-generativeai